+++ Radiant/Radiant/GameState.cpp
@@ -132,7 +132,7 @@ void GameState::Init()
 			float r = (rand() % 255) / 255.0f;
 			float g = (rand() % 255) / 255.0f;
 			float b = (rand() % 255) / 255.0f;
-			_builder->Light()->BindPointLight(e, XMFLOAT3((float)x, 2.0f, y), 1.0f, XMFLOAT3(r, g, b), 5.0);
+			_builder->Light()->BindPointLight(e, XMFLOAT3((float)x, 2.0f, (float)y), 1.0f, XMFLOAT3(r, g, b), 5.0f);
 			_builder->Transform()->SetPosition(e, XMVectorSet((float)x, 0.5f, (float)y, 0.0f));
 			i--;
 		}
